Update script.js
First off, want to say the code is nice and organized and it was easy for me to find where stuff was! 

My only real feedback is the logic for the event listener you're adding to the card. I think this is a rather large and unwieldy block of code, that should be broken up into multiple functions.  You can use event bubbling to catch the card click, then use the card's "state" to make decisions, like whether it should be visually flipped or what class to add to it. At the top I noticed you have items. You can think of cards themselves as objects too, that have properties you can manipulate.